--- a/cmd/parquetgen/gen/template.go
+++ b/cmd/parquetgen/gen/template.go
@@ -13,6 +13,7 @@ import (
 	"strings"
 	"encoding/binary"
 
+	"github.com/valyala/bytebufferpool"
 	"github.com/parsyl/parquet"
 	sch "github.com/parsyl/parquet/schema"
 	{{.Import}}
@@ -28,6 +29,8 @@ const (
 	compressionGzip         compression = 2
 	compressionUnknown      compression = -1
 )
+
+var buffpool = bytebufferpool.Pool{}
 
 // ParquetWriter reprents a row group
 type ParquetWriter struct {