--- a/cmd/parquetgen/gen/template_required.go
+++ b/cmd/parquetgen/gen/template_required.go
@@ -35,9 +35,11 @@ func (f *{{.FieldType}}) Read(r io.ReadSeeker, pg parquet.Page) error {
 }
 
 func (f *{{.FieldType}}) Write(w io.Writer, meta *parquet.Metadata) error {
-	var buf bytes.Buffer
+	buf := buffpool.Get()
+	defer buffpool.Put(buf)
+
 	for _, v := range f.vals {
-		if err := binary.Write(&buf, binary.LittleEndian, v); err != nil {
+		if err := binary.Write(buf, binary.LittleEndian, v); err != nil {
 			return err
 		}
 	}