--- a/cmd/parquetgen/gen/template_required.go
+++ b/cmd/parquetgen/gen/template_required.go
@@ -89,10 +89,10 @@ func (i *{{.TypeName}}stats) add(val {{.TypeName}}) {
 	}
 }
 
-func (f *{{.TypeName}}stats) bytes(val {{.TypeName}}) []byte {
-	var buf bytes.Buffer
-	binary.Write(&buf, binary.LittleEndian, val)
-	return buf.Bytes()
+func (f *{{.TypeName}}stats) bytes(v {{.TypeName}}) []byte {
+	bs := make([]byte, {{byteSize .}})
+	binary.LittleEndian.{{ putFunc . }}(bs, {{ uintFunc . }})
+	return bs
 }
 
 func (f *{{.TypeName}}stats) NullCount() *int64 {