--- a/cmd/parquetgen/gen/template_string.go
+++ b/cmd/parquetgen/gen/template_string.go
@@ -30,7 +30,7 @@ func (f *StringField) Write(w io.Writer, meta *parquet.Metadata) error {
 		if err := binary.Write(buf, binary.LittleEndian, int32(len(s))); err != nil {
 			return err
 		}
-		buf.Write([]byte(s))
+		buf.WriteString(s)
 	}
 
 	return f.DoWrite(w, meta, buf.Bytes(), len(f.vals), f.stats)