--- a/cmd/parquetgen/gen/template_string_optional.go
+++ b/cmd/parquetgen/gen/template_string_optional.go
@@ -44,10 +44,11 @@ func (f *StringOptionalField) Scan(r *{{.StructType}}) {
 }
 
 func (f *StringOptionalField) Write(w io.Writer, meta *parquet.Metadata) error {
-	buf := bytes.Buffer{}
+	buf := buffpool.Get()
+	defer buffpool.Put(buf)
 
 	for _, s := range f.vals {
-		if err := binary.Write(&buf, binary.LittleEndian, int32(len(s))); err != nil {
+		if err := binary.Write(buf, binary.LittleEndian, int32(len(s))); err != nil {
 			return err
 		}
 		buf.Write([]byte(s))