--- a/go.mod
+++ b/go.mod
@@ -7,4 +7,5 @@ require (
 	github.com/bxcodec/faker/v3 v3.6.0
 	github.com/golang/snappy v0.0.2
 	github.com/stretchr/testify v1.7.0
+	github.com/valyala/bytebufferpool v1.0.0 // indirect
 )