--- a/go.mod
+++ b/go.mod
@@ -4,6 +4,7 @@ go 1.13
 
 require (
 	github.com/apache/thrift v0.13.0
+	github.com/bxcodec/faker/v3 v3.6.0
 	github.com/golang/snappy v0.0.2
 	github.com/stretchr/testify v1.7.0
 )