… cmd/parquetgen/dremel/testcases/doc/generated.go
… cmd/parquetgen/dremel/testcases/person/generated.go
… cmd/parquetgen/dremel/testcases/repetition/generated.go
… go.sum
… parquet_generated_test.go
… performance/base/parquet.go