This PR contains a few self contained improvements, focused on reducing the memory usage of the parquet encoding.
each single optimization is done in a dedicated commit, with a commit message that starts with `OPT:`, so it might make more sense to review those commit individually 

to be able to compare with the baseline performance, a `parquet.go` file was generated in `performance/base/`, before applying any optimizations.
the optimized code was generated in `performance/parquet.go`, and during the benchmarks, both were used to compare the performance, and to make sure that the binary representation of the data wasn't changed by mistake.

all of the optimizations (except the last one), were done in the generated code, so it didn't effect the base version.
here is the comparison (without the last commit):
```
benchmark               iter      time/iter      bytes alloc              allocs
---------               ----      ---------      -----------              ------
BenchmarkWrite/base-8     13   865.17 ms/op   672684318 B/op   8938970 allocs/op
BenchmarkWrite/opt-8      22   518.54 ms/op   256610628 B/op   4318254 allocs/op
```

here are the results after applying the final commit as well (so there is some improvement for the base version as well):
```
benchmark               iter      time/iter      bytes alloc              allocs
---------               ----      ---------      -----------              ------
BenchmarkWrite/base-8     13   845.20 ms/op   619640619 B/op   8929125 allocs/op
BenchmarkWrite/opt-8      22   505.66 ms/op   202235762 B/op   4310258 allocs/op
```