@lev112 2021-08-17
@cswank would you be interested in this kind of change?