@cswank 2021-08-17
I'm still looking to make sure I understand this PR, but in the meantime would you mind running 'go generate' in the following dirs:

./
./cmd/parquetgen/dremel/testcases/doc/
./cmd/parquetgen/dremel/testcases/repetition/
./cmd/parquetgen/dremel/testcases/person/

So all the tests are using these changes?