@cswank 2021-08-17
Thanks @lev112.  I like the performace improvements` and the passing tests make me feel good, so... merging.